[form-builder] Block editor - better focusPath change's scroll into view logic
